Update PathSpec -> PathTest
